New backups (17+2r): Restoring data from the login screen (#3003)
* New backups: Restoring data from the login screen

This is work in progress. For time being I commented out encryption.

* bringing back kotlinx-coroutines-core
* TODO comments and excluding old coroutines library version

+++ app/src/main/kotlin/com/waz/zclient/feature/backup/messages/LikesBackUpDataSource.kt
@@ -12,7 +12,7 @@ import java.io.File
 data class LikesBackUpModel(
     val messageId: String = String.empty(),
     val userId: String = String.empty(),
-    val timeStamp: Int = 0,
+    val timeStamp: Long = 0,
     val action: Int = 0
 )
